refactor : user 패키지 정리
