+++ src/main/java/com/mopl/api/domain/user/controller/UserController.java
@@ -9,7 +9,7 @@ import com.mopl.api.domain.user.dto.response.UserDto;
 import com.mopl.api.domain.user.dto.request.UserLockUpdateRequest;
 import com.mopl.api.domain.user.dto.request.UserRoleUpdateRequest;
 import com.mopl.api.domain.user.service.UserService;
-import com.mopl.api.global.config.security.CustomUserDetails;
+import com.mopl.api.global.config.security.claim.CustomUserDetails;
 import jakarta.validation.Valid;
 import java.util.UUID;
 import lombok.RequiredArgsConstructor;
@@ -23,7 +23,6 @@ import org.springframework.web.bind.annotation.PatchMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestHeader;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestPart;
 import org.springframework.web.bind.annotation.RestController;
