+++ src/main/java/com/mopl/api/domain/user/exception/auth/detail/JwtInformationNotFoundException.java
@@ -1,4 +1,7 @@
-package com.mopl.api.domain.user.exception.auth;
+package com.mopl.api.domain.user.exception.auth.detail;
+
+import com.mopl.api.domain.user.exception.auth.AuthErrorCode;
+import com.mopl.api.domain.user.exception.auth.AuthException;
 
 public class JwtInformationNotFoundException extends AuthException {
