+++ src/main/java/com/mopl/api/domain/user/exception/user/detail/DuplicateEmailException.java
@@ -1,4 +1,7 @@
-package com.mopl.api.domain.user.exception.user;
+package com.mopl.api.domain.user.exception.user.detail;
+
+import com.mopl.api.domain.user.exception.user.UserErrorCode;
+import com.mopl.api.domain.user.exception.user.UserException;
 
 public class DuplicateEmailException extends UserException {
