+++ src/main/java/com/mopl/api/domain/user/service/UserServiceImpl.java
@@ -10,8 +10,8 @@ import com.mopl.api.domain.user.dto.response.CursorResponseUserDto;
 import com.mopl.api.domain.user.dto.response.UserDto;
 import com.mopl.api.domain.user.entity.User;
 import com.mopl.api.domain.user.entity.UserRole;
-import com.mopl.api.domain.user.exception.detail.UserNotFoundException;
-import com.mopl.api.domain.user.exception.user.DuplicateEmailException;
+import com.mopl.api.domain.user.exception.user.detail.UserNotFoundException;
+import com.mopl.api.domain.user.exception.user.detail.DuplicateEmailException;
 import com.mopl.api.domain.user.mapper.UserMapper;
 import com.mopl.api.domain.user.repository.UserRepository;
 import com.mopl.api.global.config.security.jwt.JwtRegistry;
